requirements.txt
fastapi
uvicorn
langchain
chromadb
sentence-transformers
openai
pypdf
langchain-community
python-dotenv
huggingface-hub
